Modify issues query to only pull body or labels if needed for matcher
This commit modifies the Matchinator to communicate if it needs an
issue's body or labels based on the configured matchers. If this data is
not needed, then it is not pulled by the GitHubinator. This will save on
the number of queries that are performed as well as the amount of data
that is pulled from the GitHub API, with the trade-off that the body
text and labels for an issue may not show up in a list command.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>