github: Fix bug resulting in incorrect author on items
This commit introduces a fix for a bug that causes items to have
incorrect authors. The author was set using a pointer to a GitHubAuthor
struct, rather than a unique value. The value for the pointer could be
changed after the item was produced, causing the author for the item to
be changed incorrectly.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>