github: Pull body of issue on match
This commit modified the githubinator to pull an issue's body when it is
matched, and for when the body is required for a body regex. This will
allow for users to still see the body of an issue even if they aren't
using a body regex, without requiring all issues to have their body text
pulled.

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>